Pulled HttpClientextension from Tokenservice to separate class. Minor cleanup.
